initial type checking script
